Animation Remapper has better error messages and no longer allows duplicate bindings
Can't have duplicate pending changes for a single animation clip in the
Animation Remapper. This would technically merge these values, but it
was also possible to lose keyframes based on these merges so it is no
longer supported.

Additionally the Animation Remapper has a new look and feel for errors
and notes that make them fit within the bounds of the window better.

Also simplified the colors for Animation Remapper. Now the only colors
to appear are red = broken and yellow = pending change.